+++ src/Components/Navigation/Navigation.js
@@ -3,7 +3,7 @@ import { NavLink } from "react-router-dom";
 
 import logo from "../../Assets/images/logo.svg"
 import "./Navigation.css";
-import Button from "../Button/Butt"
+import Button from "../Button/Button"
 
 
  class Navigation extends Component{
