+++ src/Components/TextAndFormMainPage/TextAndForm.css
@@ -1,6 +1,5 @@
 .text-and-form{	
-	width: 70%;
-	margin-left: 130px;
+	margin-left: 200px;
 	margin-top: 250px; 
 }
 .text-and-form h2{
@@ -18,9 +17,13 @@
     line-height: 1.75;
     color: #3b3f48;
 }
-@media (min-width: 1024px) and (max-width: 1348px) {
+@media (min-width: 1131px) and (max-width: 1400px) {
 .text-and-form{	
-	width: 100%;
-	margin-left: -55px;
+	margin-left: 50px;
 }
 }
+@media (min-width: 1024px) and (max-width: 1130px) {
+	.text-and-form{	
+		margin-left: -30px;
+	}
+	}
